small fixes
git-svn-id: https://svn.grrrr.org/ext/trunk@1086 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flthr.cpp
@@ -625,7 +625,7 @@ bool flext::ThrCond::TimedWait(double ftm)
 	ftime(&tmb);
 #endif
 	tm.tv_nsec = tmb.millitm*1000000;
-	tm.tv_sec = tmb.time; 
+	tm.tv_sec = (long)tmb.time; 
 #else // POSIX
 #if 0 // find out when the following is defined
 	clock_gettime(CLOCK_REALTIME,tm);
